@@ -1,10 +1,9 @@
 # This file is dual licensed under the terms of the Apache License, Version
 # 2.0, and the BSD License. See the LICENSE file in the root of this repository
 # for complete details.
-from __future__ import absolute_import, division, print_function
 
 
-class InfinityType(object):
+class InfinityType:
     def __repr__(self):
         # type: () -> str
         return "Infinity"
@@ -45,7 +44,7 @@ class InfinityType(object):
 Infinity = InfinityType()
 
 
-class NegativeInfinityType(object):
+class NegativeInfinityType:
     def __repr__(self):
         # type: () -> str
         return "-Infinity"
