@@ -1,7 +1,6 @@
 # This file is dual licensed under the terms of the Apache License, Version
 # 2.0, and the BSD License. See the LICENSE file in the root of this repository
 # for complete details.
-from __future__ import absolute_import, division, print_function
 
 import re
 
@@ -60,7 +59,7 @@ def canonicalize_version(version):
 
     # Epoch
     if version.epoch != 0:
-        parts.append("{0}!".format(version.epoch))
+        parts.append(f"{version.epoch}!")
 
     # Release segment
     # NB: This strips trailing '.0's to normalize
@@ -72,15 +71,15 @@ def canonicalize_version(version):
 
     # Post-release
     if version.post is not None:
-        parts.append(".post{0}".format(version.post))
+        parts.append(f".post{version.post}")
 
     # Development release
     if version.dev is not None:
-        parts.append(".dev{0}".format(version.dev))
+        parts.append(f".dev{version.dev}")
 
     # Local version segment
     if version.local is not None:
-        parts.append("+{0}".format(version.local))
+        parts.append(f"+{version.local}")
 
     return "".join(parts)
 
@@ -89,21 +88,21 @@ def parse_wheel_filename(filename):
     # type: (str) -> Tuple[NormalizedName, Version, BuildTag, FrozenSet[Tag]]
     if not filename.endswith(".whl"):
         raise InvalidWheelFilename(
-            "Invalid wheel filename (extension must be '.whl'): {0}".format(filename)
+            f"Invalid wheel filename (extension must be '.whl'): {filename}"
         )
 
     filename = filename[:-4]
     dashes = filename.count("-")
     if dashes not in (4, 5):
         raise InvalidWheelFilename(
-            "Invalid wheel filename (wrong number of parts): {0}".format(filename)
+            f"Invalid wheel filename (wrong number of parts): {filename}"
         )
 
     parts = filename.split("-", dashes - 2)
     name_part = parts[0]
     # See PEP 427 for the rules on escaping the project name
     if "__" in name_part or re.match(r"^[\w\d._]*$", name_part, re.UNICODE) is None:
-        raise InvalidWheelFilename("Invalid project name: {0}".format(filename))
+        raise InvalidWheelFilename(f"Invalid project name: {filename}")
     name = canonicalize_name(name_part)
     version = Version(parts[1])
     if dashes == 5:
@@ -111,7 +110,7 @@ def parse_wheel_filename(filename):
         build_match = _build_tag_regex.match(build_part)
         if build_match is None:
             raise InvalidWheelFilename(
-                "Invalid build number: {0} in '{1}'".format(build_part, filename)
+                f"Invalid build number: {build_part} in '{filename}'"
             )
         build = cast(BuildTag, (int(build_match.group(1)), build_match.group(2)))
     else:
@@ -124,14 +123,14 @@ def parse_sdist_filename(filename):
     # type: (str) -> Tuple[NormalizedName, Version]
     if not filename.endswith(".tar.gz"):
         raise InvalidSdistFilename(
-            "Invalid sdist filename (extension must be '.tar.gz'): {0}".format(filename)
+            f"Invalid sdist filename (extension must be '.tar.gz'): {filename}"
         )
 
     # We are requiring a PEP 440 version, which cannot contain dashes,
     # so we split on the last dash.
     name_part, sep, version_part = filename[:-7].rpartition("-")
     if not sep:
-        raise InvalidSdistFilename("Invalid sdist filename: {0}".format(filename))
+        raise InvalidSdistFilename(f"Invalid sdist filename: {filename}")
 
     name = canonicalize_name(name_part)
     version = Version(version_part)
